% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_association.R
\name{lin_associations}
\alias{lin_associations}
\title{Calculates the linear association between a matrix of features A and a vector y.}
\usage{
lin_associations(
  X,
  Y,
  W = NULL,
  n.min = 4,
  shrinkage = T,
  alpha = 0,
  MHC_direction = NULL
)
}
\arguments{
\item{X}{n x m numerical matrix of features (missing values are allowed).}

\item{Y}{n x p numerical matrix of responses (missing values are allowed)}

\item{W}{n x p numerical matrix of confounders (missing values are not allowed).}

\item{n.min}{Numeric value controlling the minimum degrees of freedom required for p-values.}

\item{shrinkage}{Boolean to control whether adaptive shrinkage should be applied.}

\item{alpha}{Numeric value controling the alpha value for adaptive shrinkage.}

\item{X.ind}{Boolean indicating whether X is the independent variable.}
}
\value{
A list with components:
\describe{
  \item{N}{Vector of degrees of freedom for each feature.}
  \item{rho}{Vector of uncorrected beta estimates controlled for W.}
  \item{beta}{Vector of corrected beta estimates controlled for W.}
  \item{beta.se}{Vector of standard errors of the beta estimates.}
  \item{p.val}{Vector of p-values (without adaptive shrinkage).}
  \item{q.val}{Vector of q-values (without adaptive shrinkage).}
  \item{res.table}{A table with the following columns (values calculated with adaptive shrinkage): \cr
                     betahat: estimated linear coefficient controlled for W. \cr
                     sebetahat: standard error for the estimate of beta. \cr
                     NegativeProb: posterior probabilities that beta is negative. \cr
                     PositiveProb: posterior probabilities that beta is positive \cr
                     lfsr: local and global FSR values. \cr
                     svalue: s-values.\cr
                     lfdr: local and global FDR values. \cr
                     qvalue: q-values for the effect size estimates. \cr
                     PosteriorMean: moderated effect size estimates. \cr
                     PosteriorSD: standard deviations of moderated effect size estimates. \cr
                     dep.var: dependent variables (column names of Y). \cr
                     ind.var: independent variables (column names of X).
                  }
}
}
\description{
Calculates the linear association between a matrix of features A and a vector y.
}
